requirements.txt
streamlit
openai
kerykeion
pyswisseph
lunar-python
pytz